+++ Ruddarr/Views/Movies/Releases/MovieReleasesView.swift
@@ -145,6 +145,10 @@ extension MovieReleasesView {
 
             indexersPicker
 
+            if !instance.releases.releaseGroups.isEmpty {
+                releaseGroupPicker
+            }
+
             qualityPicker
 
             if !instance.releases.languages.isEmpty {
@@ -224,6 +228,24 @@ extension MovieReleasesView {
             Label(
                 sort.indexer == .all ? String(localized: "Indexer") : sort.indexer,
                 systemImage: "building.2"
+            )
+        }
+    }
+
+    var releaseGroupPicker: some View {
+        Menu {
+            Picker("Release Group", selection: $sort.releaseGroup) {
+                Text("Any Group").tag(String.all)
+
+                ForEach(instance.releases.releaseGroups, id: \.self) { group in
+                    Text(group).tag(Optional.some(group))
+                }
+            }
+            .pickerStyle(.inline)
+        } label: {
+            Label(
+                sort.releaseGroup == .all ? String(localized: "Release Group") : sort.releaseGroup,
+                systemImage: "person.2"
             )
         }
     }