+++ Ruddarr/Views/Settings/SettingsPreferencesSection.swift
@@ -13,6 +13,7 @@ struct SettingsPreferencesSection: View {
         Section {
             tabPicker
             gridPicker
+            releasesPicker
             releaseFiltersPicker
 
             if ![.unknown, .notSubscribed].contains(subscriptionStatus) {
@@ -69,9 +70,24 @@ struct SettingsPreferencesSection: View {
                 Text(style.label)
             }
         } label: {
-            Label("Grid", systemImage: "square.grid.2x2")
+            Label("Grid", systemImage: "rectangle.grid.3x2")
                 .labelStyle(.settingsIcon)
         }.tint(.secondary)
+    }
+
+    @ViewBuilder
+    var releasesPicker: some View {
+        @Bindable var settings = settings
+
+        Picker(selection: $settings.releases) {
+            ForEach(ReleaseLayout.allCases) { value in
+                Text(value.label)
+            }
+        } label: {
+            Label("Releases", systemImage: "rectangle.grid.1x3")
+                .labelStyle(.settingsIcon)
+        }
+        .tint(.secondary)
     }
 
     @ViewBuilder